refactor: Github와 관련된 Infrastructure 모듈이 Application Layer에 의존하지 않도록 변경
