+++ src/main/java/com/woowacourse/kkogkkog/auth/application/OAuthService.java
@@ -1,8 +1,8 @@
 package com.woowacourse.kkogkkog.auth.application;
 
 import com.woowacourse.kkogkkog.auth.application.dto.request.AccessTokenRequest;
-import com.woowacourse.kkogkkog.auth.infrastructure.client.dto.response.OAuthAccessTokenResponse;
-import com.woowacourse.kkogkkog.auth.infrastructure.client.dto.response.OAuthProfileResponse;
+import com.woowacourse.kkogkkog.auth.application.dto.response.OAuthAccessTokenResponse;
+import com.woowacourse.kkogkkog.auth.application.dto.response.OAuthProfileResponse;
 import com.woowacourse.kkogkkog.auth.repository.InMemoryOAuthClientRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
