+++ src/test/java/com/woowacourse/kkogkkog/auth/application/AuthServiceTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.woowacourse.kkogkkog.annotation.IntegrationTest;
 import com.woowacourse.kkogkkog.auth.application.dto.response.AccessTokenResponse;
-import com.woowacourse.kkogkkog.auth.infrastructure.client.dto.response.OAuthProfileResponse;
+import com.woowacourse.kkogkkog.auth.application.dto.response.OAuthProfileResponse;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
